9.0.0A
Added Teel to the Foodhall
Disabled Hibani home for the time being.
Added Teal NPC
Fixed Hibani chats
Added Maple Syrup Candle (Necessary for summoning Teel)
Aggron SPO
Sath SPO
TONS OF UI CHANGES TO BELLYBOUND

This is not a full release, just something to hold on to or something.